Correctly restart EPG fetching after a call to flushEPG()
If flushEPG() was called without an argument (or with a 0,0,0
uniqueEPGKey) it called channel_data::startEPG(). This restarted
the EPG fecth, but without clearing the state of the sections being
processed. That led to the current fetch timing out, and the next
fetch occurring then only after an additional UPDATE_INTERVAL delay.

If flushEPG() was called with a specific uniqueEPGKey to flush a
single channel, no attempt was made to restart the EPG fetch, and
the channel was not cleared from channelLastUpdated.

Changed to use a channel_data::abortEPG()/channel_data::startChannel()
in both cases pair to cleanly end the current fetch for the channel
and to start a new one after ZAP_DELAY.

In the case where a single channel is flushed, clear the channel's
last update time from channelLastUpdated.

Update the comment about the last update time not being cleared.