[IceTV,EpgList,eEPGCache] Allow merging of IceTV & EIT EPG Allow merging of IceTV & EIT EPG on a channel-by-channel basis.
[eEPGCache]
Allow EIT and EPG_IMPORT EPGs to co-exist in the one EPG, with
channels that have had an EPG_IMPORT clearing their EIT EPG on the
first import and ignoring it after that.

Note the sources of a channel's EPG in EventCacheItem so that switches between
EOIT EPG and imports can be detected.

[IceTV]
Add new config variable config.plugins.icetv.merge_eit_epg (defaults
to False, so that default behaviour is unchanged).

Move setIceTVDefaults() to after the EPG is cleared, so that it can
apply the corect EPG source mask for the users settings.

In setIceTVDefaults() set config.epg.eit to the value of
config.plugins.icetv.merge_eit_epg so that the EIT EPG can be enabled
while IceTV is running.

makeChanServMap() now sets self.channel_service_map directly (but
still returns its value for backward compatibility). It also
calculates the set of all active IceTV triples in self.service_set.
This is used to detect users enabling and disabling channels on the
IceTV server side, and for EpgList to be able to detect whether a
channel EPG is from IceTV.

Detect changes in the user's selection of IceTV channels on the
server side, and adjust the EPG channel data appropriately.

When a channel is enabled on the server side, fetch the whole EPG
for that channel, rather than using the last update time. The normal
last update time is used to fetch the other channels' EPGs.

shouldProcessTimer() can now only returrn true if the EPG for the
channel is filled by IceTV. This prevents timers set on EIT EPG
channels from being sent to IceTV.

In IceTVNewUserSetup & IceTVNeedPassword, change their respective
keyboard() methods to test on the type of the config variable
instead of which instance it is when checking to see whether
starting the VirtualKeyBoard is appropriate. This is more general
than testing specific config variable instances.

[EpgList]
When mergiing of IceTV and EIT EPGs is enabled, tag channels with
IceTV EPG with the IceTV "star".