RecordTimer: Report write errors without suggesting the cause to be full disk
There are many other possible reasons for the recording to have failed.
Misleading the user with a possibly incorrect guess is not helpful.

Unfortunately, the event mechanism for signaling an error does not have
the means to pass the error code, so providing a specific error to the
user is not possible.