[KeyBindings.py] Unify codes across images (#1719)
* [KeyBindings.py] Unify codes across images

- Unify a coherent set of keyDescriptions, data and methods across images.
- PEP8 clean up the file.

This change aligns the key mappings across images and will allow for improvements to the help system.

If there are any issues with the key mappings please ensure that the rcpositions.xml and index.html files are actually correct for your remote control.  Many of these files have incorrect button placements and mappings.

* [KeyBindings.py] Incorrect config variable enabled

I am working on an update to the remote control code and accidentally left the new config variable enabled.