[UsageConfig.py] Add option to display screen path
This option enables a new feature in the updated "Screen.py" to display the screen path history of how the user got to the current screen.  This is made available via a new skin widget "ScreenPath".  The information can also be nominated to display in the current "Title" widget.

NOTE: Skins will need to be updated to add the "ScreenPath" widget to actually display the results of this setting.  The default has been set to "off" to preserve the previous absence of this feature.  The setting of "small" will cause the screen path history to be displayed in the new "ScreenPath" widget.  The setting of "large" will cause the "Title" widget to display they whole screen path (the "ScreenPath" widget will be cleared).