Refactor Screen.py and add screen path feature (#1744)
* [GUISkin.py] Delete file move content to Screen.py

This file was only ever used by Screen.py.  For efficiency and to simplify code tracing and management this file has been deleted and its contents integrated into Screen.py.

* [Screen.py] Merge in old GUISkin code

- Merge in GUISkin.py code.
- Use Python 3 friendly print syntax.
- Organise code methods.
- camelCase variables.
- Add comments with a "DEBUG:" label to pose questions or suggestions to be considered.

* [mytest.py] Adjust Screen.py variable reference

Change the variable name "Screen.global_screen" to "Screen.globalScreen" to match the updated name in Screen.py.

* [UsageConfig.py] Add option to display screen path

This option enables a new feature in the updated "Screen.py" to display the screen path history of how the user got to the current screen.  This is made available via a new skin widget "ScreenPath".  The information can also be nominated to display in the current "Title" widget.

NOTE: Skins will need to be updated to add the "ScreenPath" widget to actually display the results of this setting.  The default has been set to "off" to preserve the previous absence of this feature.  The setting of "small" will cause the screen path history to be displayed in the new "ScreenPath" widget.  The setting of "large" will cause the "Title" widget to display they whole screen path (the "ScreenPath" widget will be cleared).

* [setup.xml] Add screen path setting

This option enables a new feature in the updated "Screen.py" to display the screen path history of how the user got to the current screen.  This is made available via a new skin widget "ScreenPath".  The information can also be nominated to display in the current "Title" widget.

NOTE: Skins will need to be updated to add the "ScreenPath" widget to actually display the results of this setting.  The default has been set to "None" to preserve the previous absence of this feature.  The setting of "Small" will cause the screen path history to be displayed in the new "ScreenPath" widget.  The setting of "Large" will cause the "Title" widget to display they whole screen path (the "ScreenPath" widget will be cleared).

* [StreamingClientsInfo.py] Remove unused argument

Remove the unused "menu_path" calling argument.

* [Multibootmgr.py] Remove unused argument

Remove the unused "menu_path" calling argument.

* [ChoiceBox.py] Remove unused argument

- Remove the unused "menu_path" calling argument.
- Remove unused "menu_path_compressed" screen widget.  This functionality is now fully implemented on OpenATV using the "ScreenPath" widget.

* [Screen.py] Correct logging label