Screen / Skin updates (#1827)
* [skin.py] Correct a print statement

Correct a print statement to be ready for Python 3 that was missed in a previous change.

* [skin.py] Enable processing of templated screens

This code allows embedded screens to be a template of a screen scaled to HD resolution together with scaling data to allow that screen to be resized to the current GUI resolution.

* [skin.py] Add skin screen validation code

This option works in conjunction with a change in Screen.py to allow code to check if a screen in the skin is compatible with the code.  If the list of widgets considered mandatory are not present then the current screen will be rejected and the next option will be tested.  It is assumed that if all skin options are rejected that the embedded screen will be acceptable.

* [Screen.py] Add facility for mandatory widgets

This option works in conjunction with a change in Skin.py to allow code to check if a screen in the skin is compatible with the code.  If the list of widgets considered mandatory are not present then the current screen will be rejected and the next option will be tested.  It is assumed that if all skin options are rejected that the embedded screen will be acceptable.

* [Screen.py] Add option to suppress screen path

This option allows coders to suppress the normal screen path processing for a particular screen.  For example, some coders don't want MessageBoxes to have the full screen path.

* [Screen.py] Optimise applySkin()

- Optimise the code for applySkin() to remove code that is not needed or unused.
- Ensure that the title text of embedded screens is translated.

* [Screen.py] Add new display screen class

This new class will be the core support from creating a uniform and distinct class for managing front panel displays.