[IceTV] Add sanity checks to IceTV EPG and timer values
Check the values of:
    original_network_id, transport_stream_id & service_id in IceTV channels
    start & stop times, and duration in IceTV EPG entries
    IceTV genre (category) ids
    IceTV timer start & stop times, and duration

Change the fallback code for eit_id in convertChanShows() so that
it doesn't try to do int(None). Add similar fallback code to
processTimers so that if there is fallback, the timer event ids
will match the EPG cache event ids. This also ensures that the event
ids are always in the allowed range.

In the EPG processing, move the extraction of the title and show
descriptions to before the extraction and testing of the times &
duration so that those fields can be used in error messages.

Bump version number.