Add further syscall to fix spinner after recording
Add further syscall poll with timeout of 100 milliseconds as additional exit condition and consuming the "stop" signal SIGUSR1.

Do not move m_stop after actual read out of the bytes < 0 condition as we would not write any remaining bytes. In this case m_stop is evaluated by the while loop.