Fix memory leak in MovieList (2. try)
This memory leak occurs when listing the movies via the WebIf.
There is a cyclic reference between MovieList and eListboxPythonMultiContent
when the list build function is set. This prevents object destruction and so
the complete movie list stays in memory forever.
WebIf does not call reload, so it's save to set the function there.
Using a weakref for the method does not work as c++ cannot call it.

# Conflicts:
#	lib/python/Components/MovieList.py