Fix EPG when nothing is playing
If nothing is playing (e.g. watching a recording, timers start, stop
watching and now the original service is no longer available) the Single
EPG would not open and the Graphical EPG would not let you navigate.  In
this case get the current service from the InfoBar, restoring the EPGs.