[PliExtraInfo] Refactor PliExtraInfo to make it more flexible
Refactor PliExtraInfo so that conversions that produce multiple
conversion strings (like "All" and "TransponderInfo") are generated
from tables rather than directly in code.

Make use of this facility to add a "User" conversion where the skin
author may customise which conversions to include.

Also add "TransponderFrequencyMHz" to display the transponder in
MHz to 3 decimal places.

These facilities will be used in changes to the pre-loaded skins
that allow for more than 4 tuner state indicators to be shown in
the infobar.