Revert "[PliExtraInfo] Refactor PliExtraInfo to make it more flexible"
This reverts commit ad1fb9eea214afb215f7ad832fd6b52df5838b81.

The original commit only calculated TrensformerFrequencyMHz for DVB-T
and ATSC, because it assumes the data ftom the transponder inf is in Hz.
That data is in kHz for DVB-C and DVB-S.

There are also some conversions and some other minor improvements
that are in OpenViX and OpenAtV that I want to add before re-applying
the refactor.