Fix bug #752: Shutdown timers can shut down active recordings
[RecordTimer]

Add a from_time parameter to getNextZapTime(), getNextRecordingTime()
and getNextRecordingTimeOld() to allow the next timer search to be
done from a specified time.

[PowerTimer]

Add a from_time parameter to recordingsActive() to pass to calls
of getNextZapTime() and getNextRecordingTime().

Use the new parameter to check for the next timer using the currently
active PowerTimer's self.next, the activation time, rather than the
current time.

This will mean that a recording timer that starts at the same time
as the power timer to be considered to be not in the past.