[RecordTimer,PowerTimer] Clean up testing for running recordings
Move recordingsActive() from being a function in PowerTimer to being
a method in RecordTimer.

Add useStillRecording parameter and matching code so that
RecordTimer.recordingsActive() can also be used for the test for
active recordings for shutdown on "auto" and "go to deep standby"
recording after events, and use that in that test in
RecordTimerEntry.activate().

Adjust calls of recordingsActive() in PowerTimer to correspond with
the function having been moved. Reword the comment before the calls
to reflect the new calling structure.