Fix bug #756: Startup wizard scan region and final welcome screens broken
[IniTerrestrialLocation]

In class IniTerrestrialLocation, use the widget name "description"
for the screen description when it is being skinned as a setup
screen (default skin name IniTerrestrialLocation) and the widget
name "text" when it is being skinned as a Wizard (skin name
IniTerrestrialLocationWizard).

Use conditionals in the embedded skin so that the correct widget
skin definition is used in either case.

In class IniEndWizard, revert the description widget name back to
text, because it's a screen that's only ever skinned as a Wizard.