Fix bug #737: File system check always uses fsck.ext4, even when inappropriate
[Harddisk]

Add function getRealFsType() to get a partition's actual file system
type in cases where it can't be determined from the mount table
(e.g. where it's a FUSE mount).

Use getRealFsType() in getProcMounts() to use the actual file system
on FUSE mounts.

Add fsckOpts to Harddisk to list the program name and parameters
used to check different file system types.

Add attribute fs_type to Harddisk instances to hold the (first
mounted) partition's filesystem type.

In createCheckJob(), replace explicit tests on whether ext4 is
supported on the system with a lookup of fsckOpts to find the
appropriate fsck program and its parameters.

If no appropriate fsck program can be found for a partition,
createCheckJob() returns None.

[HarddiskSetup]

Allow for createCheckJob() (or any other Job creator in Harddisk)
returning None. Only a generic error can be given at this point in
the code.