Fix bug #737: Improve HDD selection in storage format/check
[Harddisk]

Copy the list of user-friendly named for filesystem types from
Screens.Action into Partition, remove the fake-up of the translation
of fuseblk to NTFS, and add some more filesysteem types. Make the
user-friendly filesystem type name translatable. Add
getFsUserFriendlyType() to Partition to convert a system filesystem
type into a user-friendly one, and add getFsUserFriendlyType( to
Harddisk to return the user-friendly filesystem type of the HDD's
first mounted partition.

Add a method checkIsSupported() to indicate whether Harddisk has
an options table entry for the filesystem and that its fsck program
is available.

[HarddiskSetup]

When the menu list is constructed, call method isSelectable() on
each item to allow subclasses to control whether list entries can
be selected. Add an always-True isSelectable() HarddiskSelection,
and add an isSelectable() to HarddiskFsckSelection thal only allows
checking to be run on HDDs that have the necessary checkers.

Adjust the menu list to the width and item height of its skin. Add
a second entry in each list element for the filesystem type. Remove
unused font=1 definition, and forced item height from the code.

Remove unnecessary and incorrect translations in
SubHarddiskMenuEntryComponent to allow passing None as the item
argument to make an entry non-seectable. Grey out unselectable
items.

Change the "Job can not be run" popup in hddConfirmed() to be headed
by "Internal error", since that error should not be able to occur.