Fix marked text offset
The offset for marked text was stored in the context.  This caused an
incorrect offset for screens with multiple labels.  In particular, the
virtual keyboard input line could shift after selecting another key.
Store the offset directly in eLabel & eListboxPythonMultiContent and
pass it as a parameter to renderText.  As a consequence of this change,
the space used to indicate the end of a config string has become a
non-breaking space (as space is removed by right alignment).