Update it.po
The previous translation is wrong and confusing. 
The word "usage" must be translated to "utilizzo".
The word "GUI" can never be translated to "sistema", which actually means "system"! It could be translated literally to "interfaccia utente", but can also be leaved as "GUI" which is a known english term.
Moreover there is already another voice under the same menu, called "sistema". So it is confusing.