[NetworkSetup.py] Harden NetworkPassword code
This change restores the previous, more secure, code but adds some further protections and improvements.
- Force the "passwd" program to run in US English mode so that the prompts can be interpreted by the Python code.
- Add logic to ensure that the interactions with "passwd" are correct and complete.
- Force a timeout if the interactions with "passwd" do not go as expected.
- Correct the analysis of the status code returned from the "execute()" method call.  This status is *NOT* the exit status of the command being run!  This is the status of the attempt to launch the command being run.
- Move the analysis of the exit status of the "passwd" command run to where the exit value is actually presented.
- Add some more logging to show what is going on in the UI.

The issue with the first attempt was that some systems run in various languages other than US English.  The "passwd" program is one of the few that will actually change its UI language.  This was causing the programmed interactions from NetworkPassword to not be found.  There is now a nasty hack in the "keySave()" method to force the "passwd" program to run in US English.  This hack can and should be removed when all Enigma2 plugins stop inappropriately changing the active shell language on a whim.