[MessageBox.py] Change "self.list" initialization
This changes returns the initialization of the "self.list" variable to an empty list rather than None to protect skins with applet embedded Python code that uses "self.list" without checking that it exists or is a valid list.  The change improves compatibility with older screens that test the length of self.list without first checking that it exists as a list.

NOTE: All skin applets with embedded Python code should include try/except around all the code to ensure that errors in the skin Python do not crash the skin or Enigma2.