[skin.py] Remove superfluous reloadWindoStyles() call
This call was added to ensure that the current windowstyle is always activated.  The fact that window styles was repeatedly activated was not considered an issue.

Recent work on openATV has been targeted at optimising the performance and code.  With this in mind we would like to review this code and look to optimising it.

There are some new changes and enhancements coming to the window styles and we would like to see skin processing optimised to improve the speed of the UI and user experience.  If any skin developers or users detect any screen or skin changes resulting from this change please alert us as a matter of urgency.  Please only report any issues *DIRECTLY* caused by this one change!