[Menu.py] Add menu images
This change adds menu images to all "Menu.py" based screens.  This is done in a style similar to how images were added to "Setup.py" screens.  The images are defined in a "menus" block within the skin.  Each "menu" tag in the block contains a "key" attribute which is the "entryID" of the menu entry to which this image will be attached.  The "menu" tag also contains an "image" attribute that is the path to the menu image to be used for that menu entry.

Similar to "Setup.py" "Setup" screen the "Menu.py" "Menu" screen can have a "menuimage" widget which defines where the menu image will be displayed.  This image will change to reflect the currently selected menu entry.  If a "default" image is defined then this image will be used for all menu entries that do not have a specifically assigned image.

The new "Menu.py" also allows for a smaller version of the image to be displayed on each menu entry line.  This is similar to the old "Info" screen which was retired because this new system was in development.  How the menu entries are displayed is controlled by the new "Menu entry style" setting in the "OSD Settings" setup screen.  This option is a replacement for the old "Show menu numbers" setting.  The reason for the change is that menu entries can now have the text alone, the text with a number, the text with an image or the text with an image and a number.

Note that all the setting controls for "Menu.py" are now all located together in the "OSD Settings" screen.