[config.py] Enhance getResolvedKey() method
This change adds an optional new argument to the getResolvedKey() method.  This option, "silent", allows code to call getResolvedKey() and, if the key is missing/invalid, return None rather than an empty string.  The silent option also suppresses the error log for the key error.

This option has been added for code, like the ConfigEntryTest converter, that may access invalid config items.  The option suppresses all the unwanted log entries.  The option also allows the calling code to identify if the requested key was not found instead of just returning an empty string value.

The reason for adding this option was triggered by skins wanting to know if the IceTV plugin is installed and active.  If the plugin is not installed then the required key was not available and the logs could fill with appropriate missing key errors.  This option silences those unwanted logs.