[ScreenSaver.py] Match class name to module name
This change makes the class name match the module name.  The code has also been tidied up with the imports sorted and optimized and variables camelCased.  Also the randint() function has been replaced with the base randrange() function as the lowest value for the X and Y coordinates should be 0 not 1 (the left and top screen coordinates are 0 not 1).

The class name change is also reflected in "InfoBarGenerics.py".

The code defaults to using the newer screen name of "ScreenSaver" but will fall back to using the previous "Screensaver" screen name for skins that aren't updated.