[config.py] Remove units argument from ConfigSelectionNumber
This change was made in conjunction with the addition of the new module TranslationData.py.  This was added as an aid to ensuring all strings are added to the translation harvest. After reviewing this code and discussing the feature with other developers we agree that this implementation will not deliver the results we want. It creates an opportunity for parts of the code to get out of sync.

We will now change ConfigSelectionNumber instances that require units into ConfigSelection instances with appropriate choices lists.