[skin.py] Suppress confusing skin error logs
Some skin attributes are optional and need not be specified to achieve the default value.  This change sets the default values in the dictionary get() functions to suppress the "[Skin] Error: The value 'None' is not a valid integer!" errors that were being generated when the attribute was not defined.  (Even though the message was displayed the correct value was being set.)

Remove the specified default value of "None" as this is not required as this is the default.