Add InfoBarShowHide hook
This change allows plugins to be notified when the InfoBarShowHide class is loaded.  The notification will include the reason (1=Started, 0=Stopped), the instance of the InfoBar being started/stopped and the type of InfoBar that is being started/stopped.

[InfobarGenerics.py]
- Add the plugin hooks for the WHERE_INFOBARLOADED event.

[Plugin.py]
- Add the WHERE_INFOBARLOADED type of event.
- Tidy up the code.  Sort the imports.  Optimize the code.

(cherry picked from commit 66f389e2d535b5b78de8afd9180001ec928860ad)