[IceTV] Hide SMS help window when not wanted
Add widgets for HelpWindow to the embedded skins for IceTVNewUserSetup
& IceTVNeedPassword and set them so that the window will be off-screen.

Remove unnecessary assignment to selh["HelpWindow"] in IceTVNeedPassword.

This is to hide the SMS HelpWindow in screens that are too small
to have one.

In skins that define screens for IceTVNewUserSetup or IceTVNeedPassword
the screens should define HelpWindow widgets to position the
HelpWindow appropriately for the screen. within the screan area,
or off the physical screen.