Fix IceTV infinite loop on EPG download & various other fixes (#2618)
* [IceTV] Fix infinite loop on IceTV EPG fetch

The changes to indentation in IceTV.plugin.EPGFetcher.processShowsBatched()
in commit 938ee5 weren't correct.

This commit corrects the indents and fixes the infinite loop.

* [IceTV] Fix fetch of PVR MAC address

Use Py2/Py3 portability constructs when fetching MAC address in
IceTV.API.getMacAddress().

The original code was throwing an exception because the code assumed
str, while Py3 required/returned bytearray. The exception meant
that the function always returned a MAC address "00:00:00:00:00:00".

* [IceTV] Fix integer division

The original Py2 code used '/' for
integer / integer -> integer division.

Corrected that to use "//" for compatible Py2/Py3
integer / integer -> integer division.

* [IceTV] Fix crash on missing IceTV icons

Fix bug where, if the IceTV epgclock_icetv.png is missing from both
the current skin and the default skin, and the fillin of non-IceTV
EPG channels by the EIT EPG is enabled, then the graphical and
multi EPGs will crash if they try to display any channels with IceTV
EPG.

Ensures that attributes of EpgList.icetvicon aren't accessed if
EpgList.icetvicon is None.

* [IceTV] Hide SMS help window when not wanted

Add widgets for HelpWindow to the embedded skins for IceTVNewUserSetup
& IceTVNeedPassword and set them so that the window will be off-screen.

Remove unnecessary assignment to selh["HelpWindow"] in IceTVNeedPassword.

This is to hide the SMS HelpWindow in screens that are too small
to have one.

In skins that define screens for IceTVNewUserSetup or IceTVNeedPassword
the screens should define HelpWindow widgets to position the
HelpWindow appropriately for the screen. within the screan area,
or off the physical screen.

* [IceTV] Complete removal of IceTV DE

Complete the partial removal of the German IceTV service
done in commit 614c60b.

Put IceTV DE QR code entry back in, but commented out, and
comment out "Germany" in the IceTV server list.

* [IceTV] Clarify where IceTV is available

Make it explicit in the IceTV setup wizard that the IceTV service
is only available in the listed countries in the service selection
screen.