Fix MultiBoot for differing boot slots
This change fixes an issue on a handful of boxes, including the SF8008 family, Beyonwiz V2, Ulan and Amiko, that use different technologies for some of the boot slots.

This change also introduces a new MultiBoot debug log option to allow for collection of more information should these sorts of issues arise again.