[MediaPlayer.py] Fix playlist file selection bug
This code is part of a bigger refactor of the FileCommander / MoviePlayer / FileList code.  This small section has been brought forward to address issues being raised in GitHub and the forum that the MediaPlayer is stuck playing the wrong files (typically is always plays the last file in the playlist no matter what file is selected) when OK is pressed.