[config.py] Improve ConfigSelectionNumber
Allow for the number value to be displayed with units text if specified in the calling code.  This allows for the numbers to be displayed in a more meaningful way.  For example, instead of an option list of "1, 2, 3, 4, 5" you could display "1 Second, 2 Seconds, 3 Seconds, 4 Seconds, 5 Seconds".

NOTE: The unit strings used in the units list/tuple should be defined elsewhere for translation as this code can not translate variables!