[FileCommander] Revert TAB manipulation
Replacing TABs with a sequence of characters can cause problems with the text if the substitution string is incorrectly manipulated by users editing the text.  The original intention of this comment was to replace the tab with another *single* character that can be entered by the user.  I didn't implement this code originally because I wasn't sure of a safe character to use for the substitution.  I was thinking that we could change VirtualKeyBoard to deliver the TAB editing functionality.  This requires more planning and consideration before being implemented.

By the way, the substitution should have been for the representation of a tab "\t" and not a TAB itself.