Move "getChipSetString()" to a BoxInfo variable
This change moves the "getChipSetString()" method from the about.py module to a BoxInfo variable called "ChipsetString".  The variable is immutable as the box can't change its chipset during operation.  ;)

The new variable can be accessed with "BoxInfo.getItem("ChipsetString")".

A side benefit of this change is that the chipset is determined only once.  This can reduce file access and time taken to get the value for this variable.