[FileList.py] Improve parent name icon label (#2857)
* [FileList.py] 
- Improve parent name icon label
This change renames the name of the parent icon in the file list from "back.png" to a more obvious "parent.png".

- Add more icon options
Add even more options for different icons to be used for special directories.  The new icon options include:
- "storage.png"
- "current.png"
- "parent.png"

If any of these icons is undefined in the skin then the default "directory.png" icon will be used.