[Setup.py] Update comments to warn of pending change
This update changes the comments for the Summary screen to advise coders and skinners that the proposed widget name changes are now available and should be used.  The old widgets "SetupTitle", "SetupEntry" and "SetupValue" are now deprecated and will be removed soon.

All openATV display skins will be updated to use the new widget names.