[PluginBrowser] * Add new setting to define the physical picon install location. * Implementation of multiselect. (Use OK for select and then GREEN to start the action)
Please note!
The picon install location of picon ipk packages is always "/picon" and this will be a symlink to the physical target.
This physical target can be defined in the settings or PluginBrowser.
Multiselect inside of "Manage Plugins" has a small limitation.
You cannot mix install and remove in one hit.