[config.py] Add a new ConfigSelectionInteger class
The new ConfigSelectionInteger class is a replacement for the ConfigSelectionNumber class.  The new class uses a comparable argument list with most other classes where "default" is the first argument.  The variables "min" and "max" are renamed to avoid the functions of the same name.

The new class also features a setChoices() method to allow for the list of valid integers to be changed to reflect changed conditions.

The old ConfigSelectionNumber class is retained and mapped into the new class for compatibility.